优化启动的代码

+++ app/src/main/java/com/runningsnail/demos/common/utils/DensityUtils.java
@@ -29,7 +29,7 @@ public class DensityUtils {
      */
 
     public static int dp2px(Context context, float dpVal) {
-        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dpVal, context.getResources().getDisplayMetrics());
+        return (int)TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dpVal, context.getResources().getDisplayMetrics());
     }
 
 
@@ -37,8 +37,8 @@ public class DensityUtils {
      * sp转px
      */
 
-    public static int sp2px(Context context, float spVal) {
-        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, spVal, context.getResources().getDisplayMetrics());
+    public static float sp2px(Context context, float spVal) {
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, spVal, context.getResources().getDisplayMetrics());
     }
